:art: move payload-types.ts
