+++ src/actions/Pages/getFaqPage.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { getPayload } from "@/libs/payload";
-import { FaqPage } from "@/payload-types";
+import { FaqPage } from "@/types/payload-types";
 
 export const getFaqPage = async (): Promise<FaqPage> => {
     const payload = await getPayload();
@@ -9,6 +9,5 @@ export const getFaqPage = async (): Promise<FaqPage> => {
         slug: "faq-page",
     });
 
-    // @ts-expect-error TODO FIX THIS
     return faqPage;
 };
