+++ src/actions/getAllProjects.ts
@@ -2,7 +2,7 @@
 
 import { SLUG } from "@/libs/enums/slug";
 import { getPayload } from "@/libs/payload";
-import { Project } from "@/payload-types";
+import { Project } from "@/types/payload-types";
 
 // return all projects
 export const getAllProjects = async (): Promise<Project[]> => {
