+++ src/types/parser/ParsePayloadAboutPage.ts
@@ -1,5 +1,5 @@
 import placeholder from "@/assets/image/600x400.png";
-import { AboutPage as CmsAboutPage, Media } from "@/payload-types";
+import { AboutPage as CmsAboutPage, Media } from "@/types/payload-types";
 
 import { AboutPage } from "../pages/AboutPage";
