+++ src/types/parser/ParsePayloadHeroPage.ts
@@ -2,7 +2,7 @@ import { VariantProps } from "tailwind-variants";
 
 import { card } from "@/app/(site)/_components/WhatWeDoCard";
 import placeholder from "@/assets/image/600x400.png";
-import { HeroPage as CmsHeroPage, Partner } from "@/payload-types";
+import { HeroPage as CmsHeroPage, Partner } from "@/types/payload-types";
 
 import { HeroPage } from "../pages/HeroPage";
