+++ src/types/parser/ParsePayloadProject.ts
@@ -1,6 +1,6 @@
 import { media } from "@/libs/payload";
 import { isNullish, removeDuplicates } from "@/libs/utils";
-import type { Project as CMSProject } from "@/payload-types";
+import type { Project as CMSProject } from "@/types/payload-types";
 
 import type { Project } from "../models";
